+++ examples/tutorials/plot_xt.py
@@ -3,36 +3,67 @@
 Expected threat
 ===============
 
-This example shows how to create expected threat (xT). This was
-introduced by Karun Singh in this `great blog post <https://karun.in/blog/expected-threat.html>`_
+This example shows how to create expected threat (xT). Expected threat is a way of valuing
+the likelihood of scoring when in possession of the ball at a position on the football pitch.
+
+Expected threat is based on `markov chains <https://en.wikipedia.org/wiki/Markov_chain>`_.
+The key to modelling soccer in this way is the probability of scoring only depends on the current
+action, it is memoryless, it does not consider what happened before or after the action.
+Often in soccer this isn't a fair assumption as attacks may form quickly on the counter or due
+to pressuring the opponent high up the field.
+
+I really recommend reading through this excellent
+`blog post <https://soccermatics.medium.com/explaining-expected-threat-cbc775d97935>`_ by
+`David Sumpter (@soccermatics) <https://twitter.com/Soccermatics>`_ on the history of expected
+threat, its limitations, and extensions.
+
+The first use of markov chains to evaluate the probability of scoring was by
+`Sarah Rudd <https://twitter.com/srudd_ok>`_ in their
+`conference presentation <http://nessis.org/nessis11/rudd.pdf>`_ "a framework for tactial
+analysis and individual offensive production assessment in soccer using markov chains." Although
+not named expected threat it contained many of the ideas used here. The idea was then popularised
+and named by `Karun Singh <https://twitter.com/karun1710>`_
+in their `fantastic interactive blog post <https://karun.in/blog/expected-threat.html>`_. In this
+tutorial we model expected threat using Karun's ideas.
 """
 
-from mplsoccer import Sbopen, Pitch
-import numpy as np
-import pandas as pd
 import matplotlib.patheffects as path_effects
 import matplotlib.pyplot as plt
+import numpy as np
+import pandas as pd
+
+from mplsoccer import Sbopen, Pitch
 
 parser = Sbopen()
 pitch = Pitch(line_zorder=2)
+
+##############################################################################
+# Set up the grid
+# ---------------
+# Our first decision, is how to grid the soccer field.
+# Here we copy Karun's setup and have 16 cells in the x-direction and
+# 12 cells in the y-direction
 bins = (16, 12)  # 16 cells x 12 cells
 
 ##############################################################################
 # Get event data
 # --------------
 # Get event data from the FA Women's Super League 2019/20.
-# Here we include only regular play events, which excludes set pieces and also counter attacks.
+# Here we include only regular play events, which excludes set pieces and also counter-attacks.
 
-df_match = parser.match(competition_id=37, season_id=42)  # 87 games from FA WSL 2019/20
+# first let's get the match file which lists all the match identifiers for
+# the 87 games from the FA WSL 2019/20
+df_match = parser.match(competition_id=37, season_id=42)
 match_ids = df_match.match_id.unique()
 
+# next we create a dataframe of all the events
 all_events_df = []
 cols = ['match_id', 'id', 'type_name', 'sub_type_name', 'x', 'y', 'end_x', 'end_y', 'outcome_name']
 for match_id in match_ids:
     # get carries/ passes/ shots
-    event = parser.event(match_id)[0]
+    event = parser.event(match_id)[0]  # get the first dataframe (events) which has index = 0
     event = event.loc[event.type_name.isin(['Carry', 'Shot', 'Pass']), cols].copy()
-    
+
     # boolean columns for working out probabilities
     event['goal'] = event['outcome_name'] == 'Goal'
     event['shoot'] = event['type_name'] == 'Shot'
@@ -43,11 +74,16 @@ event = pd.concat(all_events_df)
 ##############################################################################
 # Bin the data
 # ------------
-# Here we calculate the probability of a shot, successful move (pass or carry), and goal (given a shot).
+# Here we calculate the probability of a shot,
+# successful move (pass or carry), and goal (given a shot).
+# We are averaging the boolean columns (True = 1) and (False = 0) to give us the
+# probability between zero and one.
 
-shot_probability = pitch.bin_statistic(event['x'], event['y'], values=event['shoot'], statistic='mean', bins=bins)
-move_probability = pitch.bin_statistic(event['x'], event['y'], values=event['move'], statistic='mean', bins=bins)
-goal_probability = pitch.bin_statistic(event.loc[event['shoot'], 'x'], 
+shot_probability = pitch.bin_statistic(event['x'], event['y'], values=event['shoot'],
+                                       statistic='mean', bins=bins)
+move_probability = pitch.bin_statistic(event['x'], event['y'], values=event['move'],
+                                       statistic='mean', bins=bins)
+goal_probability = pitch.bin_statistic(event.loc[event['shoot'], 'x'],
                                        event.loc[event['shoot'], 'y'],
                                        event.loc[event['shoot'], 'goal'],
                                        statistic='mean', bins=bins)
@@ -56,35 +92,40 @@ goal_probability = pitch.bin_statistic(event.loc[event['shoot'], 'x'],
 # Plot shot probability
 # ---------------------
 fig, ax = pitch.draw()
-pitch.heatmap(shot_probability, ax=ax)
+shot_heatmap = pitch.heatmap(shot_probability, ax=ax)
 
 ##############################################################################
 # Plot goal probability
 # ---------------------
 fig, ax = pitch.draw()
-pitch.heatmap(goal_probability, ax=ax)
+goal_heatmap = pitch.heatmap(goal_probability, ax=ax)
 
 ##############################################################################
 # Plot move probability
 # ---------------------
 fig, ax = pitch.draw()
-pitch.heatmap(move_probability, ax=ax)
+move_heatmap = pitch.heatmap(move_probability, ax=ax)
 
 ##############################################################################
 # Calculate the move transition matrix
 # ------------------------------------
+# The move transition matrix takes into account the success probability of carrying
+# out the transitions. It is the probability of moving the ball successfully from one grid
+# cell to another grid cell.
 
-# get a dataframe of move events and filter out events starting outside the pitch (-1 binnumber)
+# get a dataframe of move events and filter it
+# so the dataframe only contains actions inside the pitch.
 move = event[event['move']].copy()
 bin_start_locations = pitch.bin_statistic(move['x'], move['y'], bins=bins)
-move = move[np.all(bin_start_locations['binnumber'] != -1, axis=0)].copy()
+move = move[bin_start_locations['inside']].copy()
 
-# get the successful moves and filter out the events that ended outside the pitch
+# get the successful moves, which filters out the events that ended outside the pitch
+# or where not successful (null)
 bin_end_locations = pitch.bin_statistic(move['end_x'], move['end_y'], bins=bins)
-move_success = move[(np.all(bin_end_locations['binnumber'] != -1, axis=0) &
-                     move['outcome_name'].isnull())].copy()
+move_success = move[(bin_end_locations['inside']) & (move['outcome_name'].isnull())].copy()
 
-# get a dataframe of the bin start and end locations
+# get a dataframe of the successful moves
+# and the grid cells they started and ended in
 bin_success_start = pitch.bin_statistic(move_success['x'], move_success['y'], bins=bins)
 bin_success_end = pitch.bin_statistic(move_success['end_x'], move_success['end_y'], bins=bins)
 df_bin = pd.DataFrame({'x': bin_success_start['binnumber'][0],
@@ -92,20 +133,24 @@ df_bin = pd.DataFrame({'x': bin_success_start['binnumber'][0],
                        'end_x': bin_success_end['binnumber'][0],
                        'end_y': bin_success_end['binnumber'][1]})
 
-# calculate the bin counts for the succesful moves
+# calculate the bin counts for the successful moves, i.e. the number of moves between grid cells
 bin_counts = df_bin.value_counts().reset_index(name='bin_counts')
+
+# create the move_transition_matrix of shape (num_y_bins, num_x_bins, num_y_bins, num_x_bins)
+# this is the number of successful moves between grid cells.
 move_transition_matrix = np.zeros((bins[1], bins[0], bins[1], bins[0]))
 move_transition_matrix[bin_counts['y'], bin_counts['x'],
                        bin_counts['end_y'], bin_counts['end_x']] = bin_counts.bin_counts.values
 
 # and divide by the starting locations for all moves (including unsuccessful)
+# to get the probability of moving the ball successfully between grid cells
 bin_start_locations = pitch.bin_statistic(move['x'], move['y'], bins=bins)
 bin_start_locations = np.expand_dims(bin_start_locations['statistic'], (2, 3))
 move_transition_matrix = np.divide(move_transition_matrix,
                                    bin_start_locations,
                                    out=np.zeros_like(move_transition_matrix),
-                                   where=bin_start_locations!=0,
-                                  )
+                                   where=bin_start_locations != 0,
+                                   )
 
 ##############################################################################
 # Get the matrices
@@ -119,17 +164,20 @@ goal_probability_matrix = np.nan_to_num(goal_probability['statistic'])
 ##############################################################################
 # Calculate xT
 # ------------
-# Calculate xT until convergence
-
+# Calculate xT until convergence. Initially the expected threat is set to the shot probability
+# multiplied by the goal probability. This means the expected value
+# in the first step is the probability of scoring from the grid cell if the person takes a shot.
 xt = np.multiply(shot_probability_matrix, goal_probability_matrix)
 diff = 1
 iteration = 0
-while np.any(diff > 0.00001):
-    xt_copy = xt.copy()
+while np.any(diff > 0.00001):  # iterate until the differences between the old and new xT is small
+    xt_copy = xt.copy()  # keep a copy for comparing the differences
+    # calculate the new expected threat
     xt = (np.multiply(shot_probability_matrix, goal_probability_matrix) +
-      np.multiply(move_probability_matrix, 
-                  np.multiply(move_transition_matrix, np.expand_dims(xt, axis=(0, 1))).sum(axis=(2, 3)))
-     )
+          np.multiply(move_probability_matrix,
+                      np.multiply(move_transition_matrix, np.expand_dims(xt, axis=(0, 1))).sum(
+                          axis=(2, 3)))
+          )
     diff = (xt - xt_copy)
     iteration += 1
 print('Number of iterations:', iteration)
@@ -141,11 +189,23 @@ print('Number of iterations:', iteration)
 
 path_eff = [path_effects.Stroke(linewidth=1.5, foreground='black'),
             path_effects.Normal()]
-for_plotting = pitch.bin_statistic(event['x'], event['y'], bins=bins)  # new bin statistic for plotting only
+# new bin statistic for plotting xt only
+for_plotting = pitch.bin_statistic(event['x'], event['y'], bins=bins)
 for_plotting['statistic'] = xt
 fig, ax = pitch.draw(figsize=(14, 9.625))
 _ = pitch.heatmap(for_plotting, ax=ax)
-_ = pitch.label_heatmap(for_plotting, ax=ax, str_format='{:.2%}', color='white', fontsize=14, va='center', ha='center',
-                       path_effects=path_eff)
+_ = pitch.label_heatmap(for_plotting, ax=ax, str_format='{:.2%}',
+                        color='white', fontsize=14, va='center', ha='center',
+                        path_effects=path_eff)
 
 plt.show()  # If you are using a Jupyter notebook you do not need this line
+
+##############################################################################
+# TO DO
+# -----
+# Scoring events.
+
+##############################################################################
+# TO DO
+# -----
+# Improvments using expected goals and different grid cells.